<Refact> : 검색 페이지 API 분리
- 검색 페이지에서 사용되는 검색 관련 api 코드를 기존 index.js에서 searchAPI.js로 분리
- 비록 현재는 검색 관련 api 함수가 하나이지만 추후 api가 추가될 가능성을 고려하여 관리의 용이성을 위해 객체 안에 작성함.
- 파일 분리 후 테스트를 통해 정상적으로 작동하는지 확인함.

Close #201
