+++ src/api/index.js
@@ -129,20 +129,20 @@ const FetchApi = {
   //   return data;
   // },
 
-  async searchUser(keyword, token) {
-    const response = await fetch(
-      `${BASE_URL}/user/searchuser/?keyword=${keyword}`,
-      {
-        method: 'GET',
-        headers: {
-          Authorization: `Bearer ${token}`,
-          'Content-type': 'application/json',
-        },
-      },
-    );
-    const data = await response.json();
-    return data;
-  },
+  // async searchUser(keyword, token) {
+  //   const response = await fetch(
+  //     `${BASE_URL}/user/searchuser/?keyword=${keyword}`,
+  //     {
+  //       method: 'GET',
+  //       headers: {
+  //         Authorization: `Bearer ${token}`,
+  //         'Content-type': 'application/json',
+  //       },
+  //     },
+  //   );
+  //   const data = await response.json();
+  //   return data;
+  // },
 
   // async loadFeed(token) {
   //   const response = await fetch(`${BASE_URL}/post/feed`, {
