+++ src/pages/SearchUser/SearchUser.jsx
@@ -1,9 +1,9 @@
 import { useContext, useState, useEffect, useRef } from 'react';
 import { AuthContext } from '../../context/context';
-import FetchApi from '../../api';
 import Nav from '../../components/Nav/Nav';
 import TabMenu from '../../components/common/TabMenu/TabMenu';
 import UserListItem from '../../components/UserListItem/UserListItem';
+import searchAPI from '../../api/searchAPI';
 import * as S from './StyledSearchUser';
 
 const SearchUser = () => {
@@ -38,7 +38,7 @@ const SearchUser = () => {
   useEffect(() => {
     if (keyword) {
       const getUserData = async () => {
-        const data = await FetchApi.searchUser(keyword, user.token);
+        const data = await searchAPI.searchUser(keyword, user.token);
         setIndex(15);
         setPieceUserData([]);
         setUserData(data);
